db clean test

+++ database/migrations/2019_04_04_190525_create_pictures_table.php
@@ -21,7 +21,7 @@ class CreatePicturesTable extends Migration
             $table->text('author');
             $table->text('website')->nullable();
 
-            $table->unsignedBigInteger('bird_id')->nullable();
+            $table->unsignedInteger('bird_id')->nullable();
 
             $table->timestamps();
             $table->softDeletes();
